requirements.txt
PyQt5>=5.15.0
pynput>=1.7.0
pyinstaller>=6.0.0
pyobjc-framework-ApplicationServices>=10.0;sys_platform=="darwin"
